# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "COZE-DISCORD-PROXY 代理服务",
        "title": "COZE-DISCORD-PROXY",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
...
        "/v1/chat/completions": {
            "post": {
                "description": "发送消息-openai",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openai"
                ],
                "summary": "发送消息-openai",
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.OpenAIChatCompletionRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "out-time",
                        "name": "out-time",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/model.OpenAIChatCompletionResponse"
                        }
                    }
                }
            }
        },
        "/v1/images/generations": {
            "post": {
                "description": "图片生成-openai",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openai"
                ],
                "summary": "图片生成-openai",
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.OpenAIImagesGenerationRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "out-time",
                        "name": "out-time",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/model.OpenAIImagesGenerationResponse"
                        }
                    }
                }
            }
        },
        "/v1/models": {
            "get": {
                "description": "模型列表-openai",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "openai"
                ],
                "summary": "模型列表-openai",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "schema": {
                            "$ref": "#/definitions/model.OpenaiModelListResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.OpenAIChatCompletionRequest": {
            "type": "object",
            "properties": {
                "channelId": {
                    "type": "string"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.OpenAIChatMessage"
                    }
                },
                "model": {
                    "type": "string"
                },
                "stream": {
                    "type": "boolean"
                }
            }
        },
        "model.OpenAIChatCompletionResponse": {
            "type": "object",
            "properties": {
                "choices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.OpenAIChoice"
                    }
                },
                "created": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "object": {
                    "type": "string"
                },
                "suggestions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "system_fingerprint": {
                    "type": "string"
                },
                "usage": {
                    "$ref": "#/definitions/model.OpenAIUsage"
                }
            }
        },
        "model.OpenAIChatMessage": {
            "type": "object",
            "properties": {
                "content": {},
                "role": {
                    "type": "string"
                }
            }
        },
        "model.OpenAIChoice": {
            "type": "object",
            "properties": {
                "delta": {
                    "$ref": "#/definitions/model.OpenAIDelta"
                },
                "finish_reason": {
                    "type": "string"
                },
                "index": {
                    "type": "integer"
                },
                "logprobs": {
                    "type": "string"
                },
                "message": {
                    "$ref": "#/definitions/model.OpenAIMessage"
                }
            }
        },
        "model.OpenAIDelta": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                }
            }
        },
        "model.OpenAIImagesGenerationDataResponse": {
            "type": "object",
            "properties": {
                "b64_json": {
                    "type": "string"
                },
                "revised_prompt": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "model.OpenAIImagesGenerationRequest": {
            "type": "object",
            "properties": {
                "channelId": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "prompt": {
                    "type": "string"
                },
                "response_format": {
                    "type": "string"
                }
            }
        },
        "model.OpenAIImagesGenerationResponse": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "integer"
                },
                "dailyLimit": {
                    "type": "boolean"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.OpenAIImagesGenerationDataResponse"
                    }
                },
                "suggestions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "model.OpenAIMessage": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "model.OpenAIUsage": {
            "type": "object",
            "properties": {
                "completion_tokens": {
                    "type": "integer"
                },
                "prompt_tokens": {
                    "type": "integer"
                },
                "total_tokens": {
                    "type": "integer"
                }
            }
        },
        "model.OpenaiModelListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.OpenaiModelResponse"
                    }
                },
                "object": {
                    "type": "string"
                }
            }
        },
        "model.OpenaiModelResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "object": {
                    "type": "string"
                }
            }
        },
...
    }
}